+++ 103 + 309/Project/Source/Can_HDX.h
@@ -12,12 +12,10 @@ UINT8 Can_IsBusy(void);
 void Can_PrepareSleep(void);
 
 #if PROJECT_CFG_DEBUG_MONITOR_ENABLE
-void Can_GetDebugSnapshot(uint8_t *bus_active, uint8_t *power_on,
-                          uint8_t *bus_off,  uint8_t *no_ack_cnt,
-                          uint8_t *tx_queue, uint8_t *probe,
-                          uint8_t *rtc_svc,  uint16_t *esr,
-                          uint16_t *tx_ok,   uint16_t *tx_fail,
-                          uint16_t *boff_in, uint16_t *boff_out);
+void Can_GetDebugSnapshot(uint8_t *power_on,
+                          uint8_t *bus_off,
+                          uint8_t *tx_queue,
+                          uint16_t *esr);
 #endif
 
 #endif